refactor(streaming): extract streaming turn runtime into dedicated file
Moves the `StreamingTurnRuntime` struct and its supporting `StreamingDraftBuffer` actor from `AgentRuntime.swift` to `StreamingTurnRuntime.swift`. This improves code organization and modularity by giving the streaming turn logic its own dedicated file.
